# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: iGoTrend SaaS - Influencer Marketing Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /invites:
    get:
      operationId: listMyInvites
      tags: [invites]
      summary: List creator's campaign invitations
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [pending, active, completed, declined, null]
      responses:
        "200":
          description: Invites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CampaignInvite"
  /invites/{id}/accept:
    post:
      operationId: acceptInvite
      tags: [invites]
      summary: Accept a campaign invitation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accepted
  /invites/{id}/decline:
    post:
      operationId: declineInvite
      tags: [invites]
      summary: Decline a campaign invitation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Declined

  # ─── Submissions ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    CreatorProfile:
      type: object
      required: [id, firstName, lastName, userName, badge]
      properties:
        id: { type: integer }
        firstName: { type: string }
        lastName: { type: string }
        userName: { type: string }
        email: { type: ["string", "null"] }
        badge: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        contentCategoryNames: { type: ["string", "null"] }
        creatorCategoryNames: { type: ["string", "null"] }
        instagramProfile: { type: ["string", "null"] }
        facebookProfile: { type: ["string", "null"] }
        twitterProfile: { type: ["string", "null"] }
        youtubeProfile: { type: ["string", "null"] }
        tiktokProfile: { type: ["string", "null"] }
        snapchatProfile: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        totalReach: { type: integer }
        totalEngagement: { type: integer }
        avgRating: { type: ["number", "null"] }
        campaignsCompleted: { type: integer }
        instagramDayPostPrice: { type: ["integer", "null"] }
        instagramWeekPostPrice: { type: ["integer", "null"] }
        tiktokDayPostPrice: { type: ["integer", "null"] }
        tiktokWeekPostPrice: { type: ["integer", "null"] }
        youtubeDayPostPrice: { type: ["integer", "null"] }
        youtubeWeekPostPrice: { type: ["integer", "null"] }
        gems: { type: integer }
...
    Campaign:
      type: object
      required: [id, name, sponsor, status, startDate, endDate, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        sponsor: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string, enum: [pending, active, completed, declined] }
        type: { type: string, enum: [influencer, content_creator] }
        campaignDuration: { type: string, enum: [day, weekly] }
        startDate: { type: string }
        endDate: { type: string }
        noOfCreators: { type: integer }
        estimatedBudget: { type: ["number", "null"] }
        coverImageUrl: { type: ["string", "null"] }
        invitesCount: { type: integer }
        submissionsCount: { type: integer }
        createdAt: { type: string }
...
    CampaignInvite:
      type: object
      required: [id, campaignId, creatorId, status, createdAt]
      properties:
        id: { type: integer }
        campaignId: { type: integer }
        creatorId: { type: integer }
        status: { type: string, enum: [pending, active, completed, declined] }
        campaign:
          $ref: "#/components/schemas/Campaign"
        creator:
          $ref: "#/components/schemas/CreatorProfile"
        estimatedPayout: { type: ["number", "null"] }
        createdAt: { type: string }

    # ─── Submission ───────────────────────────────────────────────────────────
...